Move `EventStream` into `.event` package.
