+++ server/src/main/java/io/spine/server/aggregate/Aggregate.java
@@ -39,9 +39,9 @@ import io.spine.server.command.CommandHandlerMethod;
 import io.spine.server.command.CommandHandlingEntity;
 import io.spine.server.entity.EventPlayer;
 import io.spine.server.entity.EventPlayingEntity;
-import io.spine.server.entity.EventStream;
 import io.spine.server.event.EventFactory;
 import io.spine.server.event.EventReactorMethod;
+import io.spine.server.event.EventStream;
 import io.spine.server.model.Model;
 import io.spine.server.rejection.RejectionReactorMethod;
 import io.spine.validate.ValidatingBuilder;
