+++ server/src/main/java/io/spine/server/entity/TransactionalEntity.java
@@ -25,6 +25,7 @@ import io.spine.annotation.Internal;
 import io.spine.core.Event;
 import io.spine.core.Version;
 import io.spine.reflect.GenericTypeIndex;
+import io.spine.server.event.EventStream;
 import io.spine.validate.ValidatingBuilder;
 import io.spine.validate.ValidatingBuilders;
