+++ server/src/main/java/io/spine/server/projection/Projection.java
@@ -26,8 +26,8 @@ import io.spine.core.EventContext;
 import io.spine.core.EventEnvelope;
 import io.spine.server.entity.EventPlayer;
 import io.spine.server.entity.EventPlayingEntity;
-import io.spine.server.entity.EventStream;
 import io.spine.server.entity.TransactionalEntity;
+import io.spine.server.event.EventStream;
 import io.spine.server.event.EventSubscriberMethod;
 import io.spine.server.model.Model;
 import io.spine.validate.ValidatingBuilder;
