+++ server/src/test/java/io/spine/server/aggregate/AggregateShould.java
@@ -46,8 +46,8 @@ import io.spine.server.aggregate.given.aggregate.TestAggregateRepository;
 import io.spine.server.aggregate.given.aggregate.UserAggregate;
 import io.spine.server.commandbus.CommandBus;
 import io.spine.server.commandbus.DuplicateCommandException;
-import io.spine.server.entity.EventStream;
 import io.spine.server.entity.InvalidEntityStateException;
+import io.spine.server.event.EventStream;
 import io.spine.server.model.Model;
 import io.spine.server.model.ModelTests;
 import io.spine.test.TimeTests;
