+++ server/src/test/java/io/spine/server/projection/ProjectionShould.java
@@ -31,8 +31,8 @@ import io.spine.core.Version;
 import io.spine.core.Versions;
 import io.spine.protobuf.TypeConverter;
 import io.spine.server.command.TestEventFactory;
-import io.spine.server.entity.EventStream;
 import io.spine.server.entity.given.Given;
+import io.spine.server.event.EventStream;
 import io.spine.validate.StringValueVBuilder;
 import org.junit.Before;
 import org.junit.Test;
