+++ server/src/test/java/io/spine/server/projection/ProjectionTransactionShould.java
@@ -25,11 +25,11 @@ import io.spine.core.Event;
 import io.spine.core.Subscribe;
 import io.spine.core.Version;
 import io.spine.core.Versions;
-import io.spine.server.entity.EventStream;
 import io.spine.server.entity.ThrowingValidatingBuilder;
 import io.spine.server.entity.Transaction;
 import io.spine.server.entity.TransactionListener;
 import io.spine.server.entity.TransactionShould;
+import io.spine.server.event.EventStream;
 import io.spine.test.projection.Project;
 import io.spine.test.projection.ProjectId;
 import io.spine.test.projection.event.PrjProjectCreated;
